🐛 Fix build issue

+++ windows/internal/ZIMPluginMethodHandler.cpp
@@ -2844,7 +2844,7 @@ void ZIMPluginMethodHandler::queryGroupApplicationList(flutter::EncodableMap& ar
                                                unsigned long long nextFlag, const ZIMError &errorInfo){
         if (errorInfo.code == 0) {
             FTMap retMap;
-            retMap[FTValue("nextFlag")] = FTValue((int64_t)nextFlag);
+            retMap[FTValue("nextFlag")] = FTValue((int32_t)nextFlag);
             retMap[FTValue("applicationList")] = ZIMPluginConverter::cnvZIMGroupApplicationInfoToArray(applicationList);
             sharedPtrResult->Success(retMap);
         } else {
